requirements.txt
langchain
langchain-openai
langchain-mongodb
langgraph==1.0.7
pymongo
kiwipiepy
tqdm
python-dotenv
pandas
beautifulsoup4
requests
lxml
aiohttp
openai
motor
sentence-transformers
transformers
torch
streamlit>=1.41.0
altair>=5.0.0
httpx
uvloop
jinja2
pydantic
langsmith
PyYAML
numpy
nbstripout
